refactor(ui/add-product-form): refactored create product form page into separate hook and component
The form that is used to create products was refactored. The logic that writes the users input to
the database was extracted into its own reusable hook. The ui for the form was then extracted into
its own component: ProductForm. The product form will be reused for updating products in addition to
creating products.