feat(ui/edit-product-inventory): added a form to edit product details
A modal form has been added with the product details auto-filled when the pencil icon is clicked on
any of the products in the inventory. Updating any of the fields and then submitting the form will
cause the product to be update both in the UI and within the database.