feat(ui/product-image-upload): added a partially working feature to upload product images
The functionality for product image uploads was partially completed. As of right now, the images
show up in the product browsing page when the product edit is saved. However, when the seller comes
back to edit the product, the images they uploaded would not be present. A working fix might be made
in the next commit.